requirements.txt
numpy==1.17.1
pandas
tensorflow-cpu
keras
pillow
flask
werkzeug
gunicorn